+++ NEWS.md
@@ -9,6 +9,9 @@ feature metadata variable columns and the results columns
 * Study packages that includes plots using base graphics now import the graphics
 package
 
+* Fix bug when querying model-specific barcode metadata. Affected both
+`getBarcodes()` and `getBarcodeData()` (reported by Paul Nordlund)
+
 # 0.6.0
 
 * New function `startApp()` to start the app running on local machine